+++ src/dsr_feature_eng_ml/models/__init__.py
@@ -1,13 +1,17 @@
 """Model specification base exports."""
 
 from dsr_feature_eng_ml.models.model_specification import (
-    ModelSpecification,
+    ClassificationModelParams,
     ModelParams,
+    ModelSpecification,
+    RegressionModelParams,
     ScikitModel,
 )
 
 __all__ = [
     "ModelSpecification",
     "ModelParams",
+    "ClassificationModelParams",
+    "RegressionModelParams",
     "ScikitModel",
 ]