+++ tests/evaluation/test_auditor.py
@@ -2,7 +2,6 @@ import dataclasses
 
 import pandas as pd
 import pytest
-
 from dsr_feature_eng_ml.enums import (
     BalancingStrategy,
     ModelType,
@@ -24,8 +23,8 @@ from dsr_feature_eng_ml.evaluation.schema import (
 )
 from dsr_feature_eng_ml.models.lasso_regression import LassoParams, LassoRegression
 from dsr_feature_eng_ml.models.random_forest import (
-    RandomForestParams,
     RandomForestRegressorModel,
+    RandomForestRegressorParams,
 )
 
 
@@ -173,8 +172,8 @@ def populated_summary(mini_taxi_df):
         scoring=ScoringMetric.R2,
         n_jobs=1,
         n_iter=10,
-        model_params=RandomForestParams.create_default(
-            task_type=TaskType.REGRESSION, scoring=ScoringMetric.R2, random_state=75
+        model_params=RandomForestRegressorParams(
+            scoring=ScoringMetric.R2, random_state=75
         ),
     )
     # 2. Create a Losing Model (Lasso)